+++ .gitignore.txt
@@ -1,0 +1,2 @@
+# .gitignore
+__pycache__/
